+++ saleor/graphql/app/types.py
@@ -8,7 +8,7 @@ from ...app.types import AppExtensionHttpMethod, AppExtensionTarget
 from ...core.exceptions import PermissionDenied
 from ...core.jwt import JWT_THIRDPARTY_ACCESS_TYPE
 from ...core.utils import build_absolute_uri
-from ...permission.auth_filters import AuthorizationFilters
+from ...permission.auth_filters import AuthorizationFilters, is_staff_user
 from ...permission.enums import AppPermission
 from ...permission.utils import message_one_of_permissions_required
 from ...thumbnail import PIL_IDENTIFIER_TO_MIME_TYPE
@@ -235,20 +235,27 @@ class AppExtension(AppManifestExtension, ModelObjectType[models.AppExtension]):
 
     @staticmethod
     @app_promise_callback
-    def resolve_app(root, info: ResolveInfo, app):
-        app_id = None
-        if app and app.id == root.app_id:
-            app_id = root.app_id
-        else:
-            requestor = get_user_or_app_from_context(info.context)
-            if requestor and requestor.has_perm(AppPermission.MANAGE_APPS):
-                app_id = root.app_id
+    def resolve_app(root, info: ResolveInfo, app_requestor):
+        # Resolve if app from context is the same as requested app
+        if app_requestor and app_requestor.id == root.app_id:
+            return AppByIdLoader(info.context).load(root.app_id)
 
-        if not app_id:
-            raise PermissionDenied(
-                permissions=[AppPermission.MANAGE_APPS, AuthorizationFilters.OWNER]
-            )
-        return AppByIdLoader(info.context).load(app_id)
+        # Reject if app from context is different from requested app
+        if app_requestor and app_requestor.id != root.app_id:
+            raise PermissionDenied(permissions=[AuthorizationFilters.OWNER])
+
+        # Resolve app if a user from the context is authenticated
+        # At this point, it's always user, because we exit for app in checks above
+        maybe_user = get_user_or_app_from_context(info.context)
+
+        # Allow staff users to access app data, no MANAGE_APPS needed
+        if maybe_user and is_staff_user(info.context):
+            return AppByIdLoader(info.context).load(root.app_id)
+
+        # If none of the conditions are met, reject with permission denied
+        raise PermissionDenied(
+            permissions=[AppPermission.MANAGE_APPS, AuthorizationFilters.OWNER]
+        )
 
     @staticmethod
     def resolve_permissions(root: models.AppExtension, _info: ResolveInfo):